--- dev/PyRFM/Enumerations/KernelPriors.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class KernelPriors(Enum):
-    LogNormals = 0
-    InverseGammas = 1
