--- dev/PyRFM/Enumerations/ObservationModels.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class ObservationModels(Enum):
-    Gaussian = 0
-    Logit = 1
-    Poisson = 2
